+++ src/lib.rs
@@ -313,17 +313,13 @@ mod tests {
 
   use test::Bencher;
 
-  use winit::event_loop::EventLoopBuilder;
   use winit::platform::x11::EventLoopBuilderExtX11 as _;
 
 
   /// Benchmark the performance of the rendering path.
   #[bench]
   fn bench_render(b: &mut Bencher) {
-    let event_loop = EventLoopBuilder::new()
-      .with_any_thread(true)
-      .build()
-      .unwrap();
+    let event_loop = EventLoop::builder().with_any_thread(true).build().unwrap();
     let mut window = Window::new(&event_loop).unwrap();
 
     let (phys_w, phys_h) = window.size();