+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Made initial AI state configurable via configuration
+- Updated `glutin` dependency to `0.31.0`
 - Updated `winit` dependency to `0.29.2`